// osa10/rate-repository-api/osa9/flights/src/types.ts
import z from 'zod';
import { newEntrySchema } from './utils';

export enum Weather {
  Sunny = 'sunny',
  Rainy = 'rainy',
  Cloudy = 'cloudy',
  Stormy = 'stormy',
  Windy = 'windy',
}

//export type Visibility = 'great' | 'good' | 'ok' | 'poor';
export enum Visibility {
  Great = 'great',
  Good = 'good',
  Ok = 'ok',
  Poor = 'poor',
}

/*export interface DiaryEntry {
    id: number;
    date: string;
    weather: Weather;
    visibility: Visibility;
    comment?: string;
  }*/

export type NonSensitiveDiaryEntry = Omit<DiaryEntry, 'comment'>;

export type NewDiaryEntry = z.infer<typeof newEntrySchema>; 

export interface DiaryEntry extends NewDiaryEntry {
  id: number;
}
//export type NewDiaryEntry = Omit<DiaryEntry, 'id'>;